Add tests for unusal arg names in exported fns/methods
Also reorganize integration tests for fns and methods into their own modules.
